requirements.txt
flask==3.0.0
torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0
anthropic>=0.25.0
python-multipart>=0.0.6
werkzeug>=3.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
safetensors>=0.4.0
pdf2image>=1.16.0
huggingface_hub>=0.19.0
accelerate>=0.25.0
# Enhanced name detection dependencies
tokenizers>=0.15.0
datasets>=2.14.0 